Organize widget.js into files
