+++ .gitignore
@@ -1,3 +1,4 @@
-src/genstudio/dist.js
+**/*compiled.js
 *.ipynb
 *.slides.html
+meta.json
