+++ package.json
@@ -18,7 +18,8 @@
     "react-dom": "^18.3.1"
   },
   "scripts":{
-    "build": "npx esbuild --bundle --format=esm src/genstudio/widget.js --outfile=src/genstudio/dist.js",
-    "watch": "npx esbuild --bundle --watch --format=esm src/genstudio/widget.js --outfile=src/genstudio/dist.js"
+    "build": "npx esbuild --bundle --minify --format=esm src/genstudio/js/widget.js --outfile=src/genstudio/js/widget_compiled.js",
+    "watch": "npx esbuild --bundle --watch --format=esm src/genstudio/js/widget.js --outfile=src/genstudio/js/widget_compiled.js",
+    "inspect": "npx esbuild --bundle --minify --format=esm src/genstudio/js/widget.js --metafile=meta.json --outfile=src/genstudio/js/widget_compiled.js && open https://esbuild.github.io/analyze/"
   }
 }
