+++ src/genstudio/js/context.js
@@ -1,0 +1,2 @@
+import {createContext} from "react"
+export const WidthContext = createContext();
